Commit all work done so far
