+++ misago/templates/misago/admin/attachments/list.html
@@ -1,5 +1,5 @@
 {% extends "misago/admin/generic/list.html" %}
-{% load i18n misago_capture misago_forms %}
+{% load i18n misago_capture misago_admin_forms %}
 
 
 {% block table-header %}
