+++ misago/templates/misago/admin/categoryroles/categoryroles.html
@@ -1,5 +1,5 @@
 {% extends "misago/admin/generic/list.html" %}
-{% load crispy_forms_field i18n %}
+{% load misago_admin_forms i18n %}
 
 
 {% block title %}
@@ -36,7 +36,7 @@
           {{ form.role }}
         </td>
         <td>
-          {% crispy_field form.category_role %}
+          {% form_field form.category_role %}
         </td>
       </tr>
       {% endfor %}
