+++ misago/templates/misago/admin/categoryroles/rolecategories.html
@@ -1,5 +1,5 @@
 {% extends "misago/admin/generic/list.html" %}
-{% load crispy_forms_field i18n %}
+{% load misago_admin_forms i18n %}
 
 
 {% block title %}
@@ -39,7 +39,7 @@
           {{ form.category }}
         </td>
         <td>
-          {% crispy_field form.role %}
+          {% form_field form.role %}
         </td>
       </tr>
       {% endfor %}
