+++ misago/templates/misago/admin/conf/group.html
@@ -1,5 +1,5 @@
 {% extends "misago/admin/conf/index.html" %}
-{% load i18n misago_forms %}
+{% load i18n misago_admin_forms %}
 
 
 {% block title %}{% trans active_group.name %} | {{ block.super }}{% endblock %}
