+++ misago/templates/misago/admin/datadownloads/list.html
@@ -1,5 +1,5 @@
 {% extends "misago/admin/generic/list.html" %}
-{% load i18n misago_avatars misago_forms %}
+{% load i18n misago_avatars misago_admin_forms %}
 
 
 {% block page-actions %}
