+++ misago/templates/misago/admin/users/delete.html
@@ -1,5 +1,5 @@
 {% extends "misago/admin/generic/form.html" %}
-{% load i18n misago_avatars misago_capture misago_forms %}
+{% load i18n misago_avatars misago_capture misago_admin_forms %}
 
 
 {% block title %}
