+++ misago/templates/misago/admin/warnings/form.html
@@ -1,5 +1,5 @@
 {% extends "misago/admin/generic/form.html" %}
-{% load i18n misago_forms %}
+{% load i18n misago_admin_forms %}
 
 
 {% block title %}
